The View3d.project method no longer allocates Point2d objects.
The allocation is now upon the caller, who can choose a more efficient
way of managing the points. In the case of XS3D nothing is actually
gained from this, but this technique is used by StarCat (unreleased at
this time) to eliminate all memory allocations which should improve
performance noticeably.